api: items/sync: always set sync_token to latest ID
Our convoluted logic was not necessary after all. The only requirement
is that `sync_token` records the latest state of the system after the
last sync. Since we use IDs as sort of a "timestamp", we can just mimic
what the Go implementation does here and always set `sync_token` equal
to the maximum ID (corresponding to the current timestamp in Go impl).

`cursor_token` is only necessary when we are not able to fetch
everything up to the latest. It has nothing to do with `sync_token` and
we should not combine the logic of the two at all.